+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/KopyCommandLineProcessor.kt
@@ -1,8 +1,10 @@
 package com.javiersc.kotlin.kopy.compiler
 
+import com.javiersc.kotlin.kopy.args.KopyVisibility
 import com.javiersc.kotlin.kopy.compiler.KopyCompilerProjectData.Group
 import com.javiersc.kotlin.kopy.compiler.KopyCompilerProjectData.Name
 import org.jetbrains.kotlin.compiler.plugin.AbstractCliOption
+import org.jetbrains.kotlin.compiler.plugin.CliOption
 import org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
 import org.jetbrains.kotlin.config.CompilerConfiguration
 
@@ -12,12 +14,12 @@ public class KopyCommandLineProcessor : CommandLineProcessor {
 
     override val pluginOptions: Collection<AbstractCliOption> =
         listOf(
-            // CliOption(
-            //     optionName = ...,
-            //     valueDescription = ...,
-            //     description = ...,
-            //     required = true,
-            // ),
+            CliOption(
+                optionName = KopyVisibility.NAME,
+                valueDescription = KopyVisibility.DESCRIPTION,
+                description = KopyVisibility.DESCRIPTION,
+                required = true,
+            ),
         )
 
     override fun processOption(
@@ -25,13 +27,7 @@ public class KopyCommandLineProcessor : CommandLineProcessor {
         value: String,
         configuration: CompilerConfiguration
     ) {
-        // when (option.optionName) {
-        //     ... -> {
-        //         configuration.put(..., value)
-        //     }
-        //     else -> {
-        //         error("Unexpected config option: ${option.optionName}")
-        //     }
-        // }
+        val kopyVisibility: KopyVisibility = KopyVisibility.from(value)
+        configuration.put(KopyKey.Visibility, kopyVisibility)
     }
 }