+++ kopy-gradle-plugin/main/kotlin/com/javiersc/kotlin/kopy/gradle/plugin/KopyExtension.kt
@@ -1,7 +1,9 @@
 package com.javiersc.kotlin.kopy.gradle.plugin
 
+import com.javiersc.kotlin.kopy.args.KopyVisibility
 import javax.inject.Inject
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Property
 import org.gradle.api.provider.ProviderFactory
 
 public open class KopyExtension
@@ -10,6 +12,9 @@ constructor(
     private val providers: ProviderFactory,
     objects: ObjectFactory,
 ) {
+
+    public val visibility: Property<KopyVisibility> =
+        objects.property(KopyVisibility::class.java).convention(KopyVisibility.Auto)
 
     public companion object {
         public const val NAME: String = "kopy"